api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymilvus==2.4.6
psycopg2-binary==2.9.9
sentence-transformers==3.0.1
torch==2.3.1

# Compat pins to avoid marshmallow v4 breakage
marshmallow==3.20.2
apispec==5.2.2
marshmallow-enum==1.5.1
